+++ README.md
@@ -1,0 +1,8 @@
+# Search time increases
+## Move ordering
+List adds each technique in sequence
+
+Alpha-beta + iterative deepen: 1180ms
+Transposition table (size 20): 768ms
+Simple sorting:                57ms
+MVv+LVA:                       33ms
