+++ src/movegen/attack_table.rs
@@ -511,7 +511,7 @@ pub const ATTACK_TABLE_SIZE: usize = 88507;
     table
 }
 // Inited from above function
-pub static ATTACKS: [u64; 88507] = [
+pub const ATTACKS: [u64; 88507] = [
     0x000040A040000000, 0x000040A040400000, 0x000040A040000000, 0x000040A040400000, 
     0x000040A040000000, 0x000040A040400000, 0x000040A040000000, 0x000040A040400000, 
     0x000040A040000000, 0x000040A040400000, 0x000040A040000000, 0x000040A040400000,
